set finalizers on Map and Program
Leaking a Map or Program currently also leaks the kernel space resource. For a long running process
this can be lethal. Use the runtime.SetFinalizer machinery to handle things a little bit more
gracefully. This also means that it's not safe to unmarshal into an existing Map anymore, but
that was probably a bad idea from the start.