Editor: Replace movsrc with movimm
The BPF verifier resets the bounds of registers in the case of
`MOV rN = rN`, which causes other instructions to be rejected as out of
bounds / invald. After rewriting globals, we sometimes end up with such
instructions.

Instead of changing the initial LOAD to load the constant value, leave
the LOAD alone and replace the MOVSRC with a MOVIMM that loads the
constant value. MOVIMM doesn't trigger the verifier bug.