add DetachMap and DetachProgram helpers
It's currently complicated to write code which reliably cleans up references
to maps or programs on errors.

    coll, err := LoadCollection("/some/file")

    // ...

    if err != nil {
        // We leak coll at this point
	return err
    }

DetachMap and DetachProgram make this a bit nicer:

    coll, err := LoadCollection("/some/file")
    // Handle err
    defer coll.Close()

    myMap := coll.DetachMap("my_map")
    defer myMap.Close()

    // ...

    if err != nil {
	return err
    }

    return coll.DetachProgram("my_program")

This has the nice side benefit that anything in coll which is not
explicitly used by the user space program is freed as soon as the enclosing
function returns.