Remove array rewriting
I never properly tested array rewriting when I introduced it,
with the result that it only works in the most narrow of circumstances.

It requires making assumptions about how clang generates code to access
the array. This is very brittle.

Remove support for now, until we can find a better way to do this.

Updates #55