editor: More reliable macro for loading constants
The 'VALUE_OF' macro used for loading rewritten constants relied on
C / clang specific hackery. This did not play particularly well with
Clang's optimizer.

Replace it with a `LOAD_CONSTANT` macro, that uses inline ASM to emit a
single load, that can easily be rewritten.
This is a bit more cumbersome to use (the macro does not expand to an
rvalue), but is more reliable / compatible with Clang's optimizer.

Does not require any change to loading / is ABI compatible with the
previous approach.