asm: fix marshalling signed jumps comparing with -1
It's perfectly valid to have an instruction like this:

    JSGTImm dst: r1 off: 4 imm: -1 ; if r1 > -1 goto PC+4

Since the offset is not -1 it does not need rewriting.
The current code is too lazy in checking for this condition,
and tries to rewrite both bpf to bpf calls and jumps in the
same block.

Separate these two cases, and add a comment so I can remember
what does what.