editor: simplify linking
The current linking code verifies that no duplicate symbols exist
in the resulting program, and tries to avoid work by linking all
libraries at once. It returns an error if a function call can
not be satisfied by any of the given libraries.

The check for duplicate symbols is not necessary, since it's performed
by asm.Instructions.Marshal later. It's also unclear how much work
is really saved.